feat: updates vul flag in projects and dependencies
The use case "UpdateVuls" was tied to the endpoint "update_vuls" for
demo purposes but it's conceived to be executed as a task.

The is_vul flag was introduced in Project and Dependency DTOs and models.
It defaults to false and we'll be updated by the use case when needed.

The "toggle_vulnerability" was added to repos to delegate the update of
a single field to the provider layer. Updating a single field can be
optimized to run faster than updating a full record even when using a
CRM. Optimization was not addressed.